Renames BetaPage as VoterRegistrationDrivePage
